+++ js/index.js
@@ -11,9 +11,9 @@ window.levelIdx = 0;
 window.turn = 0;
 window.played = false;
 window.habilidadUsadaEsteTurno = false;
-window.debugViewActive = false;
 window.selectedCardIdx = null;
 window.cronomantePortalInvestigado = null;
+window.pickerObjetivoPortal = null;
 window.juegoTerminado = false;
 window.invocationSet = 'normal';